Add tests to increase coverage in DataplexBigQueryToGcs template.
PiperOrigin-RevId: 418820756
